taipei-day-trip/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
mysql-connector-python==9.2.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-multipart==0.0.20
